Tue, Dec 6, 2022, 11:48 AM +00:00

+++ readability/readability.c
@@ -8,7 +8,7 @@ int main(void)
 
 }
 
-int averageletters(sentence)
+float averageletters(sentence)
 {
     
 }
